pyproject.toml
[tool.poetry]
name = "assessment-vokerai" 
version = "0.1.0"
description = "Backend service for order processing using FastAPI and OpenAI"
authors = ["Alireza Mounesisohi <66388192+mounesi@users.noreply.github.com>"]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
openai = "^1.58.1"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pytest = "^8.3.5"

[tool.poetry.scripts]
start = "uvicorn backend.main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"